1d type1 C++ tester going.

+++ src/finufft1d.h
@@ -3,12 +3,11 @@
 
 #include "cnufftspread.h"
 #include "twopispread.h"
-// #include "../contrib/dfftpack.h"
-#include <fftw.h>
+#include <fftw3.h>
 
-int finufft1d1(BIGINT nj,double* xj,double* cj,int iflag,double eps,int ms,
+int finufft1d1(BIGINT nj,double* xj,double* cj,int iflag,double eps,BIGINT ms,
 	       double* fk);
-int finufft1d2(BIGINT nj,double* xj,double* cj,int iflag,double eps,int ms,
+int finufft1d2(BIGINT nj,double* xj,double* cj,int iflag,double eps,BIGINT ms,
 	       double* fk);
 int finufft1d3(BIGINT nj,double* xj,double* cj,int iflag,double eps,BIGINT mi,
 	       double* fi);  // todo
